# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Restaurantes API",
        "contact": {
            "url": "",
            "email": ""
        },
        "license": {
            "name": "Deber de Programacion Web"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/pedidos": {
            "get": {
                "tags": [
                    "pedidos"
                ],
                "summary": "Lista de pedidos",
                "operationId": "listPedidos",
                "responses": {
                    "200": {
                        "description": "Un array de pedidos",
                        "headers": {
                            "x-next": {
                                "description": "A link to the next page of responses",
                                "schema": {
                                    "type": "string"
                                }
                            }
                        },
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/PedidosSchema"
                                }
                            }
                        }
                    },
                    "default": {
                        "description": "unexpected error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "pedidos"
                ],
                "summary": "Crear un pedido",
                "operationId": "createPedido",
                "requestBody": {
                    "description": "Datos del pedido",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/PedidosSchema"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Un pedido",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/PedidosSchema"
                                }
                            }
                        }
                    },
                    "default": {
                        "description": "unexpected error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/pedidos/{id}": {
            "get": {
                "tags": [
                    "pedidos"
                ],
                "summary": "Obtener pedido por id",
                "operationId": "getPedidoById",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del pedido",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Obtener un pedido por su ID",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/PedidosSchema"
                                }
                            }
                        }
                    },
                    "default": {
                        "description": "unexpected error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "pedidos"
                ],
                "summary": "Actualizar un pedido",
                "operationId": "updatePedido",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del pedido",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Datos del pedido",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/PedidosSchema"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Un pedido",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/PedidosSchema"
                                }
                            }
                        }
                    },
                    "default": {
                        "description": "unexpected error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "pedidos"
                ],
                "summary": "Eliminar un pedido",
                "operationId": "deletePedido",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del pedido",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Un pedido",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Pedidos"
                                }
                            }
                        }
                    },
                    "default": {
                        "description": "unexpected error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Error": {
                "title": "Error",
                "description": "Error",
                "properties": {
                    "code": {
                        "description": "Codigo de error",
                        "type": "integer",
                        "format": "int32",
                        "example": 400
                    },
                    "message": {
                        "description": "Mensaje de error",
                        "type": "string",
                        "example": "Error de validacion"
                    },
                    "errors": {
                        "description": "Errores",
                        "type": "array",
                        "items": {
                            "description": "Error",
                            "type": "string",
                            "example": "El campo nombre es requerido"
                        }
                    }
                },
                "type": "object"
            },
            "Pedidos": {
                "title": "Pedidos",
                "description": "Pedidos",
                "properties": {
                    "id": {
                        "description": "ID",
                        "type": "integer",
                        "example": "1"
                    },
                    "id_usuario": {
                        "description": "ID del usuario",
                        "type": "integer",
                        "example": "1"
                    },
                    "id_restaurante": {
                        "description": "ID del restaurante",
                        "type": "integer",
                        "example": "1"
                    },
                    "cantidad": {
                        "description": "Cantidad",
                        "type": "integer",
                        "example": "1"
                    },
                    "precio": {
                        "description": "Precio",
                        "type": "integer",
                        "example": "10"
                    },
                    "fecha": {
                        "description": "Fecha",
                        "type": "string",
                        "example": "2021-01-01"
                    },
                    "estado": {
                        "description": "Estado",
                        "type": "string",
                        "example": "Pendiente"
                    }
                },
                "type": "object"
            },
            "PedidosSchema": {
                "title": "Pedidos",
                "description": "Detalle del pedido",
                "properties": {
                    "id": {
                        "description": "ID",
                        "type": "integer",
                        "format": "int32",
                        "example": 1
                    },
                    "id_usuario": {
                        "description": "ID del usuario",
                        "type": "integer",
                        "format": "int32",
                        "example": 1
                    },
                    "id_restaurante": {
                        "description": "ID del restaurante",
                        "type": "integer",
                        "format": "int32",
                        "example": 1
                    },
                    "Estado": {
                        "description": "Estado del pedido",
                        "type": "string",
                        "example": "Pendiente"
                    },
                    "Fecha": {
                        "description": "Fecha del pedido",
                        "type": "date",
                        "example": "2021-01-01"
                    },
                    "Total": {
                        "description": "Total del pedido",
                        "type": "decimal",
                        "format": "decimal",
                        "example": 10
                    },
                    "detalle": {
                        "description": "Detalle del pedido",
                        "type": "array",
                        "items": {
                            "properties": {
                                "id_pedido": {
                                    "description": "ID del pedido",
                                    "type": "integer",
                                    "format": "int32",
                                    "example": 1
                                },
                                "id_plato": {
                                    "description": "ID del plato",
                                    "type": "integer",
                                    "format": "int32",
                                    "example": 1
                                },
                                "cantidad": {
                                    "description": "Cantidad",
                                    "type": "integer",
                                    "format": "int32",
                                    "example": 1
                                },
                                "total": {
                                    "description": "Total del plato",
                                    "type": "decimal",
                                    "format": "decimal",
                                    "example": 10
                                }
                            },
                            "type": "object"
                        }
                    }
                },
                "type": "object"
            },
...
        }
    }
}